Updated the thank you page
