+++ Gigue_Xamarin_Android/Gigue/Gigue.csproj
@@ -202,7 +202,6 @@
     <Compile Include="Activities\SplashScreen.cs" />
     <Compile Include="Activities\TestData.cs" />
     <Compile Include="Activities\Home.cs" />
-    <Compile Include="Activities\Thanks.cs" />
     <Compile Include="Adapters\UserData.cs" />
     <Compile Include="DataObjects\AppUser.cs" />
     <Compile Include="DataObjects\Genre.cs" />
@@ -221,7 +220,7 @@
     <Compile Include="Activities\RegistrationActivity.cs" />
     <Compile Include="Resources\Resource.designer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ThankYou.cs" />
+    <Compile Include="Activities\ThankYou.cs" />
     <Compile Include="User.cs" />
     <Compile Include="ViewModels\vmAppUser.cs" />
     <Compile Include="ViewModels\vmGenre.cs" />
