+++ Gigue_Xamarin_Android/Gigue/Resources/layout/thanks.axml
@@ -5,9 +5,33 @@
     android:layout_height="match_parent"
     android:minWidth="25px"
     android:minHeight="25px">
-    <Button
-        android:text="Button"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:id="@+id/button9" />
+  <TextView
+      android:text="Thank You!"
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content"
+      android:background="#242424"
+      android:textColor="#f2854b"
+      android:gravity="center"
+      android:textSize="18sp"
+      android:id="@+id/txtThanksHeader" />
+  <TextView
+      android:text=""
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content"
+      android:textSize="18sp"
+      android:id="@+id/txtThanksUser" />
+  <Button
+      android:text="On to Search"
+      android:id="@+id/btnThanksSearch"
+      android:background="@drawable/Login_Style_Button"
+      android:textStyle="bold"
+      android:typeface="serif"
+      android:textSize="16sp"
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content"
+      android:layout_marginTop="30dp"
+      android:layout_marginRight="100dp"
+      android:layout_marginStart="100dp"
+      android:gravity="center"
+      android:textColor="#ffffff"/>
 </LinearLayout>
