--- Gigue_Xamarin_Android/Gigue/ThankYou.cs
@@ -1,27 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-using Android.App;
-using Android.Content;
-using Android.OS;
-using Android.Runtime;
-using Android.Views;
-using Android.Widget;
-
-namespace Gigue
-{
-    [Activity (Theme = ("@android:style/Theme.NoTitleBar"))]
-    public class ThankYou : Activity
-    {
-        protected override void OnCreate(Bundle savedInstanceState)
-        {
-            base.OnCreate(savedInstanceState);
-
-            SetContentView(Resource.Layout.thanks);
-
-            // Create your application here
-        }
-    }
-}
